change double to float

+++ include/libemanelte/mhalue.h
@@ -59,7 +59,7 @@ namespace UE {
 
   std::uint64_t get_tx_prb_frequency(int prb_index, std::uint64_t freq_hz);
   
-  void set_frequencies(uint32_t carrierIndex, double carrierRxFrequencyHz, double carrierTxFrequencyHz);
+  void set_frequencies(uint32_t carrierIndex, float carrierRxFrequencyHz, float carrierTxFrequencyHz);
 
   void set_num_resource_blocks(int n_prb);
